Rename gaussian.py to fchk.py & test file
